# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: Shoe Order API
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /api/v1
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
  - application/x-www-form-urlencoded
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /orders:
    x-swagger-router-controller: order_controller
    get:
      tags: ["orders"]
      summary: Returns all created orders with the basic informations
      operationId: getAllOrders
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/GetOrdersListResponse"
    post:
      tags: ["orders"]
      summary: Create a new order
      operationId: createOrder
      parameters:
        - name: customer_id
          description: The ID of the customer who created the order
          in: formData
          required: true
          type: number
        - name: is_active
          description: The status of the order
          in: formData
          required: true
          type: boolean
      responses:
        201:
          description: Created
          schema:
            $ref: "#/definitions/PostOrderResponse"
        400:
          description: Bad Request
          schema:
            $ref: "#/definitions/PostOrderBadRequestError"
  /orders/{orderId}:
    x-swagger-router-controller: order_controller
    get:
      tags: ["orders"]
      summary: Get an order by his ID
      description: Returns an order by his ID with all basic informations
      operationId: getOrderById
      parameters:
        - name: orderId
          type: string
          in: path
          required: true
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/GetOrderResponse"
        404:
          description: Not Found
          schema:
            $ref: "#/definitions/Error"
  /orders/{orderId}/status:
    x-swagger-router-controller: order_controller
    patch:
      tags: ["orders"]
      summary: Modify the status of an order
      operationId: editOrderStatus
      parameters:
        - name: orderId
          description: Id of the order to update
          type: string
          in: path
          required: true
        - name: is_active
          description: TRUE if order is active, else FALSE
          in: formData
          required: true
          type: boolean
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/EditStatusOfOrderResponse"
        404:
          description: Not Found
          schema:
            $ref: "#/definitions/EditStatusOfOrderByIdNotFound"
...
definitions:
...
  Error:
    type: object
    properties:
      code:
        type: number
      message:
        type: string
    required:
      - code
      - message
...
  GetOrdersListResponse:
    required:
      - code
      - orders
    properties:
      code:
        type: integer
      orders:
        type: array
        items:
          type: object
          properties:
            id:
              type: integer
            customer_id:
              type: integer
            customizations_of_order_id:
              type: integer
            is_active:
              type: boolean
            created_at:
              type: string
            updated_at:
              type: string
...
  GetOrderResponse:
    required:
      - code
      - order
    properties:
      code:
        type: integer
      order:
        type: object
        properties:
          id:
            type: integer
          customer_id:
            type: integer
          is_active:
            type: boolean
          created_at:
            type: string
          updated_at:
            type: string
...
  PostOrderResponse:
    required:
      - code
      - order
    properties:
      code:
        type: number
      order:
        type: object
        properties:
          id:
            type: integer
          customer_id:
            type: integer
          is_active:
            type: boolean
          created_at:
            type: string
          updated_at:
            type: string
  PostOrderBadRequestError:
    required:
      - code
      - type
    type: object
    properties:
      code:
        type: number
      type:
        type: string
      messages:
        type: array
        items:
          type: string
      fields:
        type: array
        items:
          type: string
...
  EditStatusOfOrderResponse:
    required:
      - code
      - message
      - order
    type: object
    properties:
      code:
        type: integer
      message:
        type: string
      order:
        type: object
        properties:
          id:
            type: integer
          customer_id:
            type: integer
          is_active:
            type: boolean
          created_at:
            type: string
          updated_at:
            type: string
  EditStatusOfOrderByIdNotFound:
    required:
      - code
      - message
    type: object
    properties:
      code:
        type: integer
      message:
        type: string
...